fix(corpus): circular import
